requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.4.0
google-auth==2.5.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
mysql-connector-python==8.0.28
packaging==21.3
proto-plus==1.19.9
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pyuic5-tool==0.0.1
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
